requirements.txt
annotated-types==0.7.0
anyio==4.6.0
Authlib==1.3.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
exceptiongroup==1.2.2
fastapi==0.115.0
google-auth==2.35.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
